+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Bug Fixes
+
+* support hash rocket syntax for gem and gemspec options
+
 ## [0.8.0](https://github.com/contriboss/gemfile-go/compare/v0.7.3...v0.8.0) (2026-01-22)